+++ packages/plugin-packages/mashroom-storage-provider-mongodb/src/mongodb_client.ts
@@ -1,7 +1,11 @@
 
-import {Db, MongoClient} from 'mongodb';
+import {Db, MongoClient, Logger} from 'mongodb';
+
+import type {MongoClientOptions} from 'mongodb';
+import type {MashroomLogger} from '@mashroom/mashroom/type-definitions';
 
 let _connectionUri: string | null = null;
+let _connectionOptions: MongoClientOptions | null = null;
 let _client: MongoClient | null = null;
 let _db: Db | null = null;
 
@@ -17,12 +21,13 @@ export const close = async (): Promise<void> => {
     _db = null;
 };
 
-export const setConnectionUri = async (connectionUri: string): Promise<void> => {
+export const setConnectionUriAndOptions = async (connectionUri: string, options: MongoClientOptions): Promise<void> => {
     await close();
     _connectionUri = connectionUri;
+    _connectionOptions = options;
 };
 
-export default async (): Promise<Db> => {
+export default async (logger: MashroomLogger): Promise<Db> => {
     if (_db) {
         return _db;
     }
@@ -30,12 +35,19 @@ export default async (): Promise<Db> => {
         throw new Error('No connection URI set!');
     }
 
-    _client = new MongoClient(_connectionUri, {
-        useUnifiedTopology: true,
-        useNewUrlParser: true,
-    });
+    _client = new MongoClient(_connectionUri, _connectionOptions || {});
 
     await _client.connect();
+
+    // Redirect MongoDB logger
+    Logger.setLevel('info');
+    Logger.setCurrentLogger((msg, context) => {
+        if (context && context.type === 'info') {
+            logger.info('MongoDB:', msg);
+        } else {
+            logger.error('MongoDB:', msg);
+        }
+    });
 
     _db = _client.db();
     return _db;